+++ CHANGELOG.md
@@ -12,3 +12,5 @@
 - Added page to page loading transition
 - Added highlighting current link
 - Added check for existence of paths
+- Added axios
+- Added Redux toolkit